fix: resolve_skill_name search semantics + layer violations + budget
- Use non-anchored regex in resolve_skill_name to support "Use /..." prefix
- Update test expectation for /name form (now returns skill name per plan)
- Move server-importing test from tests/execution to tests/server
- Bump _session_model.py line budget from 370 to 385
